+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/notifier.php
@@ -126,7 +126,11 @@ return static function (ContainerConfigurator $container) {
             ->args([service('texter.transports')])
             ->tag('messenger.message_handler', ['handles' => PushMessage::class])
 
-        ->set('notifier.logger_notification_listener', NotificationLoggerListener::class)
+        ->set('notifier.notification_logger_listener', NotificationLoggerListener::class)
             ->tag('kernel.event_subscriber')
+
+        ->alias('notifier.logger_notification_listener', 'notifier.notification_logger_listener')
+            ->deprecate('symfony/framework-bundle', '6.3', 'The "%alias_id%" service is deprecated, use "notifier.notification_logger_listener" instead.')
+
     ;
 };